requirements.txt
flask
icalendar
recurring-ical-events
anthropic
apscheduler
requests
pytz
psycopg2-binary
gunicorn
python-socketio
python-engineio
openai
porcupine-client
faster-whisper
pyttsx3
python-dotenv
numpy
scipy
sounddevice
soundfile
